Samle ytelse i en bounded context
