+++ src/main/kotlin/no/nav/helse/domene/ytelse/arena/ArenaSakMapper.kt
@@ -1,21 +1,12 @@
-package no.nav.helse.domene.ytelse
+package no.nav.helse.domene.ytelse.arena
 
 import no.nav.helse.common.toLocalDate
 import no.nav.helse.domene.ytelse.domain.Kilde
 import no.nav.helse.domene.ytelse.domain.Ytelse
-import no.nav.tjeneste.virksomhet.infotrygdberegningsgrunnlag.v1.informasjon.Grunnlag
 import no.nav.tjeneste.virksomhet.meldekortutbetalingsgrunnlag.v1.informasjon.Sak
 import no.nav.tjeneste.virksomhet.meldekortutbetalingsgrunnlag.v1.informasjon.Vedtak
 
-object YtelseMapper {
-
-    fun fraInfotrygd(grunnlag: Grunnlag) =
-            Ytelse(
-                    kilde = Kilde.Infotrygd,
-                    tema = grunnlag.behandlingstema.value,
-                    fom = grunnlag.periode.fom?.toLocalDate(),
-                    tom = grunnlag.periode.tom?.toLocalDate()
-            )
+object ArenaSakMapper {
 
     fun fraArena(sak: Sak, vedtak: Vedtak) =
             Ytelse(
