+++ src/test/kotlin/no/nav/helse/domene/ytelse/infotrygd/BeregningsgrunnlagMapperTest.kt
@@ -1,8 +1,8 @@
-package no.nav.helse.domene.ytelse
+package no.nav.helse.domene.ytelse.infotrygd
 
 import no.nav.helse.common.toXmlGregorianCalendar
-import no.nav.helse.domene.ytelse.BeregningsgrunnlagMapper.toBeregningsgrunnlag
-import no.nav.helse.domene.ytelse.BeregningsgrunnlagMapper.toVedtak
+import no.nav.helse.domene.ytelse.infotrygd.BeregningsgrunnlagMapper.toBeregningsgrunnlag
+import no.nav.helse.domene.ytelse.infotrygd.BeregningsgrunnlagMapper.toVedtak
 import no.nav.helse.domene.ytelse.domain.Behandlingstema
 import no.nav.helse.domene.ytelse.domain.Beregningsgrunnlag
 import no.nav.helse.domene.ytelse.domain.Utbetalingsvedtak
@@ -10,8 +10,6 @@ import no.nav.tjeneste.virksomhet.infotrygdberegningsgrunnlag.v1.informasjon.*
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
-
-
 
 class BeregningsgrunnlagMapperTest {
